// src/core/http/generated/models/admin-billing-subscriptions/adminSubscription.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * PetVex Admin - Admin Billing Subscriptions
 * Filtered OpenAPI specification for Admin Billing Subscriptions.
 * OpenAPI spec version: 0.0.1
 */
import type { AdminSubscriptionCoupon } from './adminSubscriptionCoupon.ts';
import type { AdminTenantResource } from './adminTenantResource.ts';
import type { PlanResource } from './planResource.ts';

export interface AdminSubscription {
  id: number;
  plan_id: number;
  /** @nullable */
  status: string | null;
  /** @nullable */
  billing_cycle: string | null;
  amount: number;
  currency: string;
  gateway: string;
  /** @nullable */
  gateway_checkout_url: string | null;
  /** @nullable */
  trial_started_at: string | null;
  /** @nullable */
  trial_ends_at: string | null;
  /** @nullable */
  started_at: string | null;
  /** @nullable */
  current_period_starts_at: string | null;
  /** @nullable */
  current_period_ends_at: string | null;
  /** @nullable */
  next_billing_at: string | null;
  /** @nullable */
  canceled_at: string | null;
  /** @nullable */
  cancel_reason: string | null;
  plan?: PlanResource;
  /** @nullable */
  created_at: string | null;
  /** @nullable */
  updated_at: string | null;
  coupon: AdminSubscriptionCoupon;
  tenant_id: string;
  payment_gateway_id: string;
  billing_customer_id: string;
  gateway_customer_id: string;
  gateway_subscription_id: string;
  gateway_checkout_id: string;
  metadata: string | string[];
  tenant?: AdminTenantResource;
}