// src/core/http/generated/models/admin-billing-payments/adminPayment.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * PetVex Admin - Admin Billing Payments
 * Filtered OpenAPI specification for Admin Billing Payments.
 * OpenAPI spec version: 0.0.1
 */
import type { AdminSubscription } from './adminSubscription.ts';
import type { AdminTenantResource } from './adminTenantResource.ts';

export interface AdminPayment {
  id: number;
  /** @nullable */
  subscription_id: number | null;
  /** @nullable */
  invoice_id: number | null;
  gateway: string;
  /** @nullable */
  gateway_payment_id: string | null;
  /** @nullable */
  status: string | null;
  /** @nullable */
  status_detail: string | null;
  amount: number;
  currency: string;
  /** @nullable */
  payment_method: string | null;
  /** @nullable */
  paid_at: string | null;
  /** @nullable */
  failed_at: string | null;
  /** @nullable */
  refunded_at: string | null;
  /** @nullable */
  failure_reason: string | null;
  /** @nullable */
  created_at: string | null;
  /** @nullable */
  updated_at: string | null;
  tenant_id: string;
  payment_gateway_id: string;
  gateway_transaction_id: string;
  metadata: string | string[];
  tenant?: AdminTenantResource;
  subscription?: AdminSubscription;
}