// src/core/http/generated/models/admin-billing-invoices/adminInvoice.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * PetVex Admin - Admin Billing Invoices
 * Filtered OpenAPI specification for Admin Billing Invoices.
 * OpenAPI spec version: 0.0.1
 */
import type { AdminSubscription } from './adminSubscription.ts';
import type { AdminTenantResource } from './adminTenantResource.ts';

export interface AdminInvoice {
  id: number;
  /** @nullable */
  subscription_id: number | null;
  /** @nullable */
  number: string | null;
  /** @nullable */
  status: string | null;
  amount: number;
  currency: string;
  /** @nullable */
  due_date: string | null;
  /** @nullable */
  paid_at: string | null;
  /** @nullable */
  invoice_url: string | null;
  /** @nullable */
  created_at: string | null;
  /** @nullable */
  updated_at: string | null;
  tenant_id: string;
  payment_gateway_id: string;
  gateway_invoice_id: string;
  metadata: string | string[];
  tenant?: AdminTenantResource;
  subscription?: AdminSubscription;
}